requirementss.txt
flask~=1.1.2
gunicorn
python-binance
python-telegram-bot

pip~=22.3
zlib~=1.2.11
wheel~=0.37.1
openssl~=1.1.1h
cryptography~=38.0.1
py~=1.9.0
lxml~=4.6.1
keyring~=21.4.0
typing_extensions~=3.7.4.3
setuptools~=60.2.0
packaging~=21.3
pyparsing~=2.4.7
jinja2~=2.11.2
zipp~=3.7.0
cython~=0.29.28
future~=0.18.2
docutils~=0.16
requests~=2.18.4